backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-dateutil
chardet
numpy
pydantic
pytest
httpx
